Expose error message when .install.sh fails

+++ .install.sh
@@ -2,12 +2,20 @@
 
 set -ex
 
+VERSION=$1
+
+if [ "" == "$VERSION" ]; then
+    echo "Failed to specify version required"
+    exit 1
+fi
+
 mkdir -p h3/out
 rm -rf h3c
 git clone https://github.com/uber/h3.git h3c
 
 pushd h3c
 git pull origin master --tags
+git checkout "$VERSION"
 
 # Run CMake, installing a recent version if not found or not compatible
 {
